Main activity renamed to movie activity
