+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".ui.movies.MovieActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
