+++ app/src/main/java/com/nauman404/moviechallenge/di/builder/FragmentBuilder.kt
@@ -5,4 +5,5 @@ import dagger.Module
 @Module
 abstract class FragmentBuilder {
 
+
 }
