+++ app/src/main/java/com/nauman404/moviechallenge/ui/movies/MovieActivity.kt
@@ -1,9 +1,10 @@
-package com.nauman404.moviechallenge
+package com.nauman404.moviechallenge.ui.movies
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.nauman404.moviechallenge.R
 
-class MainActivity : AppCompatActivity() {
+class MovieActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
